added generic backflow kernel test
